move data types to own folder
