+++ appsync/aws-appsync-events/src/main/java/com/amplifyframework/aws/appsync/events/EventsChannel.kt
@@ -15,6 +15,10 @@
 package com.amplifyframework.aws.appsync.events
 
 import com.amplifyframework.aws.appsync.core.AppSyncAuthorizer
+import com.amplifyframework.aws.appsync.events.data.ChannelAuthorizers
+import com.amplifyframework.aws.appsync.events.data.EventsException
+import com.amplifyframework.aws.appsync.events.data.EventsMessage
+import com.amplifyframework.aws.appsync.events.data.PublishResult
 import kotlinx.coroutines.flow.Flow
 import kotlinx.serialization.json.JsonElement
